Improve handling of not being able to read a pid file
If a pid file was opened, but for some reason a pid could not
successfully be read, the pid used to check if a process was
running was random.

+++ keepalived/core/pidfile.c
@@ -65,10 +65,15 @@ process_running(char *pid_file)
 		return 0;
 
 	ret = fscanf(pidfile, "%d", &pid);
-	if (ret == EOF && ferror(pidfile) != 0) {
-		log_message(LOG_INFO, "Error opening pid file %s", pid_file);
+	if (ret != 1) {
+		log_message(LOG_INFO, "Error reading pid file %s", pid_file);
+		pid = 0;
 	}
 	fclose(pidfile);
+
+	/* What should we return - we don't know if it is running or not. */
+	if (!pid)
+		return 1;
 
 	/* If no process is attached to pidfile, remove it */
 	if (kill(pid, 0)) {
